Reject unrecognized arguments on the %addon line
AddonData.handle_header now accepts unknown arguments so a kickstart
containing data for an unknown addon is passed through. However, if
hello_world.py is running then addon is known, and the arguments will
not be passed through, so reject them. This raises a KickstartParseError
using the message from the previous version of addons.py.