Add a D-Bus service to perform the work for the addon
This follows the recent changes in Anaconda's internal structure:
There is a clear separation between the UI and services performing the actual
work. These parts communicate via D-Bus. Additionally there is now a clear
separation of all work into Task subclasses so that everything runs in its own
thread.

This change adds the service but it is not used yet by the UI.